+++ PluginCore/DockPanelSuite/Docking/VisibleNestedPaneCollection.cs
@@ -107,7 +107,7 @@ namespace WeifenLuo.WinFormsUI.Docking
                 Rectangle rectSplitter = rect;
                 if (status.DisplayingAlignment == DockAlignment.Left)
                 {
-                    rectThis.Width = (int)((double)rect.Width * status.DisplayingProportion) - (Measures.SplitterSize / 2);
+                    rectThis.Width = (int)(rect.Width * status.DisplayingProportion) - (Measures.SplitterSize / 2);
                     rectSplitter.X = rectThis.X + rectThis.Width;
                     rectSplitter.Width = Measures.SplitterSize;
                     rectPrev.X = rectSplitter.X + rectSplitter.Width;
@@ -115,7 +115,7 @@ namespace WeifenLuo.WinFormsUI.Docking
                 }
                 else if (status.DisplayingAlignment == DockAlignment.Right)
                 {
-                    rectPrev.Width = (rect.Width - (int)((double)rect.Width * status.DisplayingProportion)) - (Measures.SplitterSize / 2);
+                    rectPrev.Width = (rect.Width - (int)(rect.Width * status.DisplayingProportion)) - (Measures.SplitterSize / 2);
                     rectSplitter.X = rectPrev.X + rectPrev.Width;
                     rectSplitter.Width = Measures.SplitterSize;
                     rectThis.X = rectSplitter.X + rectSplitter.Width;
@@ -123,7 +123,7 @@ namespace WeifenLuo.WinFormsUI.Docking
                 }
                 else if (status.DisplayingAlignment == DockAlignment.Top)
                 {
-                    rectThis.Height = (int)((double)rect.Height * status.DisplayingProportion) - (Measures.SplitterSize / 2);
+                    rectThis.Height = (int)(rect.Height * status.DisplayingProportion) - (Measures.SplitterSize / 2);
                     rectSplitter.Y = rectThis.Y + rectThis.Height;
                     rectSplitter.Height = Measures.SplitterSize;
                     rectPrev.Y = rectSplitter.Y + rectSplitter.Height;
@@ -131,7 +131,7 @@ namespace WeifenLuo.WinFormsUI.Docking
                 }
                 else if (status.DisplayingAlignment == DockAlignment.Bottom)
                 {
-                    rectPrev.Height = (rect.Height - (int)((double)rect.Height * status.DisplayingProportion)) - (Measures.SplitterSize / 2);
+                    rectPrev.Height = (rect.Height - (int)(rect.Height * status.DisplayingProportion)) - (Measures.SplitterSize / 2);
                     rectSplitter.Y = rectPrev.Y + rectPrev.Height;
                     rectSplitter.Height = Measures.SplitterSize;
                     rectThis.Y = rectSplitter.Y + rectSplitter.Height;
